change name of the extension
