+++ app/code/Riskified/Decider/Logger/Handler.php
@@ -1,0 +1,10 @@
+<?php
+namespace Riskified\Decider\Logger;
+
+use Monolog\Logger;
+
+class Handler extends \Magento\Framework\Logger\Handler\Base
+{
+    protected $loggerType = Logger::INFO;
+    protected $fileName = '/var/log/riskified_decider.log';
+}
