+++ app/code/Riskified/Decider/Logger/Order.php
@@ -1,0 +1,6 @@
+<?php
+namespace Riskified\Decider\Logger;
+
+class Order extends \Monolog\Logger
+{
+}
