+++ app/code/Riskified/Decider/etc/module.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0"?>
+
+<config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:framework:Module/etc/module.xsd">
+    <module name="Riskified_Decider" setup_version="1.0.0">
+    </module>
+</config>
