--- app/code/Riskified/Full/Config/Source/DeclinedState.php
@@ -1,17 +1,0 @@
-<?php
-namespace Riskified\Full\Config\Source;
-
-class DeclinedState implements \Magento\Framework\Option\ArrayInterface
-{
-    /**
-     * Options getter
-     *
-     * @return array
-     */
-    public function toOptionArray()
-    {
-        return [
-            ['value' => \Magento\Sales\Model\Order::STATE_CANCELED, 'label' => __(\Magento\Sales\Model\Order::STATE_CANCELED)],
-            ['value' => \Magento\Sales\Model\Order::STATE_HOLDED, 'label' => __(\Magento\Sales\Model\Order::STATE_HOLDED)]];
-    }
-}
