--- app/code/Riskified/Full/Logger/Handler.php
@@ -1,10 +1,0 @@
-<?php
-namespace Riskified\Full\Logger;
-
-use Monolog\Logger;
-
-class Handler extends \Magento\Framework\Logger\Handler\Base
-{
-    protected $loggerType = Logger::INFO;
-    protected $fileName = '/var/log/riskified_full.log';
-}
