--- app/code/Riskified/Full/Logger/Order.php
@@ -1,6 +1,0 @@
-<?php
-namespace Riskified\Full\Logger;
-
-class Order extends \Monolog\Logger
-{
-}
