--- app/code/Riskified/Full/etc/acl.xml
@@ -1,16 +1,0 @@
-<?xml version="1.0"?>
-<config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:framework:Acl/etc/acl.xsd">
-    <acl>
-        <resources>
-            <resource id="Magento_Backend::admin">
-                <resource id="Magento_Backend::stores">
-                    <resource id="Magento_Backend::stores_settings">
-                        <resource id="Magento_Config::config">
-                            <resource id="Riskified_Full::config_settings" title="Riskified Config Settings" sortOrder="10" />
-                        </resource>
-                    </resource>
-                </resource>
-            </resource>
-        </resources>
-    </acl>
-</config>
