--- app/code/Riskified/Full/etc/frontend/events.xml
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<config>
-    <event name="sales_order_place_after">
-        <observer name="riskified_full_order_place_after_frontend" instance="Riskified\Full\Observer\OrderPlacedAfter" />
-    </event>
-</config>
