--- app/code/Riskified/Full/etc/module.xml
@@ -1,6 +1,0 @@
-<?xml version="1.0"?>
-
-<config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:framework:Module/etc/module.xsd">
-    <module name="Riskified_Full" setup_version="1.0.0">
-    </module>
-</config>
